Cleanup previous commits
