+++ client/directories.go
@@ -33,11 +33,22 @@ const (
 	logFileExt          = "teamclient"
 )
 
-// AppDir returns the teamclient directory of the app (named ~/.<app>/teamserver/client/),
-// creating the directory if needed, or logging an error event if failing to create it.
-func (tc *Client) AppDir() string {
+func (tc *Client) HomeDir() string {
 	user, _ := user.Current()
-	dir := filepath.Join(user.HomeDir, "."+tc.name, teamserverClientDir)
+	dir := filepath.Join(user.HomeDir, "."+tc.name)
+
+	err := tc.fs.MkdirAll(dir, log.DirPerm)
+	if err != nil {
+		tc.log().Errorf("cannot write to %s root dir: %s", dir, err)
+	}
+
+	return dir
+}
+
+// TeamDir returns the teamclient directory of the app (named ~/.<app>/teamclient/),
+// creating the directory if needed, or logging an error event if failing to create it.
+func (tc *Client) TeamDir() string {
+	dir := filepath.Join(tc.HomeDir(), teamserverClientDir)
 
 	err := tc.fs.MkdirAll(dir, log.DirPerm)
 	if err != nil {
@@ -50,7 +61,7 @@ func (tc *Client) AppDir() string {
 // LogsDir returns the directory of the client (~/.app/logs), creating
 // the directory if needed, or logging a fatal event if failing to create it.
 func (tc *Client) LogsDir() string {
-	logsDir := filepath.Join(tc.AppDir(), logsDirName)
+	logsDir := filepath.Join(tc.TeamDir(), logsDirName)
 
 	err := tc.fs.MkdirAll(logsDir, log.DirPerm)
 	if err != nil {
@@ -62,7 +73,7 @@ func (tc *Client) LogsDir() string {
 
 // GetConfigDir - Returns the path to the config dir.
 func (tc *Client) ConfigsDir() string {
-	rootDir, _ := filepath.Abs(tc.AppDir())
+	rootDir, _ := filepath.Abs(tc.TeamDir())
 	dir := filepath.Join(rootDir, configsDirName)
 
 	err := tc.fs.MkdirAll(dir, log.DirPerm)
