+++ server/directories.go
@@ -1,5 +1,23 @@
 package server
 
+/*
+   team - Embedded teamserver for Go programs and CLI applications
+   Copyright (C) 2023 Reeflective
+
+   This program is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
 import (
 	"fmt"
 	"os"
@@ -16,9 +34,7 @@ const (
 	logsDir       = "logs"
 )
 
-// AppDir returns the directory of the team server app (named ~/.<application>-server),
-// creating the directory if needed, or logging a fatal event if failing to create it.
-func (ts *Server) AppDir() string {
+func (ts *Server) HomeDir() string {
 	value := os.Getenv(fmt.Sprintf("%s_ROOT_DIR", strings.ToUpper(ts.name)))
 
 	var dir string
@@ -26,13 +42,26 @@ func (ts *Server) AppDir() string {
 	if !ts.opts.inMemory {
 		if len(value) == 0 {
 			user, _ := user.Current()
-			dir = filepath.Join(user.HomeDir, "."+ts.name, teamserverDir)
+			dir = filepath.Join(user.HomeDir, "."+ts.name)
 		} else {
 			dir = value
 		}
 	} else {
-		dir = filepath.Join("."+ts.name, teamserverDir)
+		dir = "." + ts.name
 	}
+
+	err := ts.fs.MkdirAll(dir, log.DirPerm)
+	if err != nil {
+		ts.log().Errorf("cannot write to %s root dir: %s", dir, err)
+	}
+
+	return dir
+}
+
+// TeamDir returns the directory of the team server app (named ~/.<app>/teamserver/),
+// creating the directory if needed, or logging a fatal event if failing to create it.
+func (ts *Server) TeamDir() string {
+	dir := path.Join(ts.HomeDir(), logsDir)
 
 	err := ts.fs.MkdirAll(dir, log.DirPerm)
 	if err != nil {
@@ -45,8 +74,7 @@ func (ts *Server) AppDir() string {
 // LogsDir returns the directory of the client (~/.app-server/logs), creating
 // the directory if needed, or logging a fatal event if failing to create it.
 func (ts *Server) LogsDir() string {
-	rootDir := ts.AppDir()
-	logDir := path.Join(rootDir, logsDir)
+	logDir := path.Join(ts.TeamDir(), logsDir)
 
 	err := ts.fs.MkdirAll(logDir, log.DirPerm)
 	if err != nil {
@@ -69,7 +97,7 @@ func (ts *Server) checkWritableFiles() error {
 	// Check home application directory.
 	// If it does not exist but we don't have write permission
 	// on /user/home, we return an error as we can't work.
-	appDirWrite, err := log.IsWritable(ts.AppDir())
+	appDirWrite, err := log.IsWritable(ts.TeamDir())
 
 	switch {
 	case os.IsNotExist(err):
